backend/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.15.1
psycopg2-binary==2.9.9
pycparser==2.22
sqlparse==0.5.0
typing_extensions==4.11.0
gunicorn
python-dotenv
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.88.0
channels==3.0.4
pillow==10.4.0